+++ cmd/branch.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/richardjennings/mygit/internal/mygit"
+	"github.com/richardjennings/mygit/pkg/mygit"
 	"github.com/spf13/cobra"
 	"log"
 	"os"
