+++ cmd/ls-files.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/richardjennings/mygit/internal/mygit"
+	"github.com/richardjennings/mygit/pkg/mygit"
 	"github.com/spf13/cobra"
 	"log"
 )
