+++ pkg/mygit/mygit.go
@@ -3,11 +3,11 @@ package mygit
 import (
 	"errors"
 	"fmt"
-	"github.com/richardjennings/mygit/internal/mygit/config"
-	"github.com/richardjennings/mygit/internal/mygit/gfs"
-	"github.com/richardjennings/mygit/internal/mygit/index"
-	"github.com/richardjennings/mygit/internal/mygit/objects"
-	"github.com/richardjennings/mygit/internal/mygit/refs"
+	"github.com/richardjennings/mygit/pkg/mygit/config"
+	"github.com/richardjennings/mygit/pkg/mygit/gfs"
+	"github.com/richardjennings/mygit/pkg/mygit/index"
+	"github.com/richardjennings/mygit/pkg/mygit/objects"
+	"github.com/richardjennings/mygit/pkg/mygit/refs"
 	"io"
 	"log"
 	"os"
