+++ pkg/mygit/mygit_test.go
@@ -3,10 +3,10 @@ package mygit
 import (
 	"bytes"
 	"fmt"
-	"github.com/richardjennings/mygit/internal/mygit/config"
-	"github.com/richardjennings/mygit/internal/mygit/gfs"
-	"github.com/richardjennings/mygit/internal/mygit/objects"
-	"github.com/richardjennings/mygit/internal/mygit/refs"
+	"github.com/richardjennings/mygit/pkg/mygit/config"
+	"github.com/richardjennings/mygit/pkg/mygit/gfs"
+	"github.com/richardjennings/mygit/pkg/mygit/objects"
+	"github.com/richardjennings/mygit/pkg/mygit/refs"
 	"github.com/stretchr/testify/assert"
 	"io/fs"
 	"os"
